created BlocklyCanvasPanel component and implemented in workspace and sandbox
