+++ client/src/views/Workspace/Workspace.js
@@ -1,27 +1,17 @@
-import React, { useEffect, useRef, useState } from "react"
-import { Link } from "react-router-dom"
-import { compileArduinoCode, getArduino, getJS, getXml, setLocalActivity } from './helpers.js'
+import React, {useEffect, useState} from "react"
 import "./Workspace.less"
-import { getActivityToolbox } from "../../Utils/requests.js"
-import { getToken } from "../../Utils/AuthRequests"
+import {getActivityToolbox} from "../../Utils/requests.js"
+import {getToken} from "../../Utils/AuthRequests"
 import Logo from "../../assets/casmm_logo.png"
 import PlaceHolderImg1 from "../../assets/science.png"
 import PlaceHolderImg2 from "../../assets/arduino.png"
 import PlaceHolderImg3 from "../../assets/maker.png"
-import { Carousel } from 'antd';
+import {Carousel} from 'antd';
+import BlocklyCanvasPanel from "../../components/BlocklyCanvasPanel/BlocklyCanvasPanel";
 
 
 export default function Workspace(props) {
-
-    const [activity, setActivity] = useState({} )
-    const [hoverXml, setHoverXml] = useState(false)
-    const [hoverJS, setHoverJS] = useState(false)
-    const [hoverArduino, setHoverArduino] = useState(false)
-    const [hoverCompile, setHoverCompile] = useState(false)
-
-    let workspaceRef = useRef(null)
-
-    const setWorkspace = () => workspaceRef.current = window.Blockly.inject('blockly-canvas', { toolbox: document.getElementById('toolbox') })
+    const [activity, setActivity] = useState({})
 
     useEffect(() => {
         const localActivity = localStorage.getItem("my-activity")
@@ -41,25 +31,7 @@ export default function Workspace(props) {
         } else {
             props.history.push('/')
         }
-
-        // clean up - removes blockly div from DOM
-        return () => {
-            if (workspaceRef.current) workspaceRef.current.dispose()
-        }
     }, [props])
-
-    useEffect(() => {
-        // once the activity state is set, set the workspace
-        if (Object.keys(activity).length && !workspaceRef.current) {
-            setWorkspace()
-            workspaceRef.current.addChangeListener(() => setLocalActivity(workspaceRef.current))
-
-            if (activity.template) {
-                let xml = window.Blockly.Xml.textToDom(activity.template)
-                window.Blockly.Xml.domToWorkspace(xml, workspaceRef.current)
-            }
-        }
-    }, [activity])
 
     return (
         <div>
@@ -75,7 +47,7 @@ export default function Workspace(props) {
                     </div>
                     <div id='bottom-container' className="flex flex-column vertical-container overflow-visible">
                         <div id="section-header">
-                                Learn about the activity...
+                            Learn about the activity...
                         </div>
                         <p id="section-text">{activity.description}</p>
                         <div id="secondary-section-header">
@@ -83,77 +55,24 @@ export default function Workspace(props) {
                         </div>
                         {/* Example implementation of image Carousel */}
                         <div id="carousel-container">
-                        <Carousel dotPosition={"left"}>
-                            <div id="diagram-container">
-                                <img id="diagram" src={PlaceHolderImg1} alt="First diagram in carousel"/>
-                            </div>
-                            <div id="diagram-container">
-                                <img id="diagram" src={PlaceHolderImg2} alt="Second diagram in carousel"/>
-                            </div>
-                            <div id="diagram-container">
-                                <img id="diagram" src={PlaceHolderImg3} alt="Third diagram in carousel"/>
-                            </div>
-                        </Carousel>
+                            <Carousel dotPosition={"left"}>
+                                <div id="diagram-container">
+                                    <img id="diagram" src={PlaceHolderImg1} alt="First diagram in carousel"/>
+                                </div>
+                                <div id="diagram-container">
+                                    <img id="diagram" src={PlaceHolderImg2} alt="Second diagram in carousel"/>
+                                </div>
+                                <div id="diagram-container">
+                                    <img id="diagram" src={PlaceHolderImg3} alt="Third diagram in carousel"/>
+                                </div>
+                            </Carousel>
                         </div>
                     </div>
                 </div>
                 <div id='horizontal-container' className="flex flex-column">
-                    <div id='top-container' className="flex flex-column vertical-container">
-                        <div id='description-container' className="flex flex-row space-between card">
-                            <Link id='link' to={"/student"} className="flex flex-column">
-                                <i className="fa fa-home" style={{"fontSize": "32px"}}/>
-                                Home
-                            </Link>
-                            <div style={{"width": "25%"}}>
-                                <div id='action-btn-container' className="flex space-between">
-                                    <i onClick={() => getXml(workspaceRef.current)} className="fas fa-code hvr-info"
-                                       onMouseEnter={() => setHoverXml(true)}
-                                       onMouseLeave={() => setHoverXml(false)}/>
-                                    {hoverXml && <div className="popup XML">Shows Xml Code</div>}
-                                    <i onClick={() => getJS(workspaceRef.current)} className="fab fa-js hvr-info"
-                                       onMouseEnter={() => setHoverJS(true)}
-                                       onMouseLeave={() => setHoverJS(false)}/>
-                                    {hoverJS && <div className="popup JS">Shows Javascript Code</div>}
-                                    <i onClick={() => getArduino(workspaceRef.current)} className="hvr-info"
-                                       onMouseEnter={() => setHoverArduino(true)}
-                                       onMouseLeave={() => setHoverArduino(false)}>A</i>
-                                    {hoverArduino && <div className="popup Arduino">Shows Arduino Code</div>}
-                                    <i onClick={() => compileArduinoCode(workspaceRef.current)}
-                                       className="fas fa-play hvr-info"
-                                       onMouseEnter={() => setHoverCompile(true)}
-                                       onMouseLeave={() => setHoverCompile(false)}/>
-                                    {hoverCompile && <div className="popup Compile">Run Program</div>}
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                    <div id='bottom-container' className="flex flex-column vertical-container overflow-visible">
-                        <div id="section-header">
-                                Program your Arduino...
-                        </div>
-                        <div id="blockly-canvas"
-                             onChange={() => setLocalActivity(workspaceRef.current)}/>
-                    </div>
+                    <BlocklyCanvasPanel activity={activity} activityType={"my-activity"}/>
                 </div>
             </div>
-
-            {/* This xml is for the blocks' menu we will provide. Here are examples on how to include categories and subcategories */}
-            <xml id="toolbox" style={{ "display": "none" }} is="Blockly workspace">
-                {
-                    // Maps out block categories
-                    activity.toolbox && activity.toolbox.map(([category, blocks]) => (
-                        <category name={category} is="Blockly category" key={category}>
-                            {
-                                // maps out blocks in category
-                                // eslint-disable-next-line
-                                blocks.map((block) => {
-                                    return <block type={block.name} is="Blockly block" key={block.name}/>
-                                })
-                            }
-                        </category>
-                    ))
-                }
-            </xml>
         </div>
     );
